channels/slack/requirements.txt
slack_sdk>=3.26.0
httpx>=0.27.0
python-dotenv>=1.0.0
